auth_service/requirements.txt
fastapi
uvicorn
sqlalchemy
asyncpg
python-jose
passlib[bcrypt]
pydantic-settings
